main_screen.py에 있는 함수 일부 다른 파일로 이동, play_screen 생성, 카드 객체 생성을 위한 class 생성
